Make core unit tests runnable again
Not very happy with their placement, but these tests rely on e-g, so
kinda make sense to put them in e-g, as opposed to core.
