+++ Cargo.toml
@@ -31,7 +31,7 @@ fixed = { version = "0.5.7", optional = true, default-features = false }
 float-cmp = "0.8.0"
 micromath = { version = "1.1.0", default-features = false }
 nalgebra = { version = "0.23.0", optional = true, default-features = false }
-embedded-graphics-core = { path=  "embedded-graphics-core", version = "0.1.0" }
+embedded-graphics-core = { path = "embedded-graphics-core", version = "0.1.0" }
 
 # criterion is not listed as a dev-dependency to work around a problem with no_std compatibility
 # and because dev-dependencies can't be optional
