+++ embedded-graphics-core/src/iterator/mod.rs
@@ -81,31 +81,3 @@ where
         pixel::Translate::new(self, offset)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::pixelcolor::BinaryColor;
-    use embedded_graphics::mock_display::MockDisplay;
-
-    #[test]
-    fn draw_pixel_iterator() {
-        let pixels = [
-            Pixel(Point::new(0, 0), BinaryColor::On),
-            Pixel(Point::new(1, 0), BinaryColor::Off),
-            Pixel(Point::new(2, 0), BinaryColor::On),
-            Pixel(Point::new(2, 1), BinaryColor::Off),
-        ];
-
-        let mut display = MockDisplay::new();
-        pixels.iter().copied().draw(&mut display).unwrap();
-
-        assert_eq!(
-            display,
-            MockDisplay::from_pattern(&[
-                "#.#", //
-                "  .", //
-            ])
-        );
-    }
-}
